microservices/investment/requirements.txt
Django==2.2.20
gunicorn==20.0.4
psycopg2-binary==2.8.4
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
django-prometheus==2.1.0
django_redis==4.11.0
django-redis-cache==3.0.0
pytest==6.2.4
ipdb==0.13.7
